WebSockets/tests/submissions/Opera/interfaces/WebSocket/events/014.html

<!doctype html>
<title>WebSockets: setting event handlers to null</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../../constants.js></script>
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
  var events = ['onclose', 'onopen', 'onerror', 'onmessage'];
  var foo = function() {}
  for (var i = 0; i < events.length; ++i) {
    ws[events[i]] = foo;
    assert_equals(ws[events[i]], foo, events[i]);
    ws[events[i]] = null;
    assert_equals(ws[events[i]], null, events[i]);
  }
});
end();
</script>
